sphinx: add new NewHopPayload constructor
In this commit, we add a new constructor to abstract away the details of
the hop payload framing from the caller. This new function accepts the
hop data, as well as optional EOB bytes. We'll then map these two pieces
of data into a single hop payload by appending the EOB bytes to the
serialized hop data.