sphinx: pad out the starting packet with random bytes
By padding out the starting packet with random bytes rather than leaving
the zeroes in tact, we patch a privacy leak that may reveal a lower
bound on the true route length to an adversarial exit node.

In order to reconcile this with our existing set of test vectors, we've
introduced a new abstraction that allows the caller to specify how they
want the starting bytes of the packet to be filled out. By default, all
callers will use the `deterministicPacketFiller`, but may also pass in the
`blankPacketFiller`, or `randomPacketFiller` variants.